pyproject.toml
[project]
name = "ai-math-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.16",
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.2",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
]
